Finally get the content builder and font handling fixed
The main bug was in how I was getting the glyph id when rendering text.
I was using the previously created character map assuming that all the
characters would be in it. However, for some fonts this was not the case
so I need to go back to the font file to determine the glyph id.